PxSEL and PxSEL2 Support
Two variables (selValue,selValue2) are defined. This values represent
the inputs which are switched to output via PxSEL,PXSEL2. This values
must be set by timer, ccu, ad-convert,... via writePortSel

Disable set in to 0 after reset. Because this are inputs and the
physical connection to other components is not affected by the reset.